requirements.txt
apache-airflow==2.8.1
pandas==2.1.4
pyarrow==14.0.2
python-dotenv==1.0.0
boto3==1.34.14
pyyaml==6.0.1
sqlalchemy
psycopg2-binary